fix algorithm + put ranking in ui

+++ src/main/java/models/bean/ClassTestData.java
@@ -13,7 +13,7 @@ public class ClassTestData  implements ITestData{
     private double ochiai;
     private double wong2;
     private double dstar;
-    private int rank;
+    private double rank;
     private boolean faulty;
     private String relativePath;
     private ArrayList<ITestData> methods;
@@ -131,11 +131,11 @@ public class ClassTestData  implements ITestData{
 
 
 
-    public int getRank() {
+    public double getRank() {
         return rank;
     }
 
-    public void setRank(int rank) {
+    public void setRank(double rank) {
         this.rank = rank;
     }
