+++ src/main/java/models/bean/MethodTestData.java
@@ -17,7 +17,7 @@ public class MethodTestData implements ITestData {
     private double ochiai;
     private double wong2;
     private double dstar;
-    private int rank;
+    private double rank;
     private boolean faulty;
     private ArrayList<ITestData> statements;
     private CloseContext closeContext;
@@ -177,11 +177,11 @@ public class MethodTestData implements ITestData {
     }
 
 
-    public int getRank() {
+    public double getRank() {
         return rank;
     }
 
-    public void setRank(int rank) {
+    public void setRank(double rank) {
         this.rank = rank;
     }
